requirements.txt
blinker==1.9.0
click==8.1.8
flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==3.0.2
mfrc522==0.0.7
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
RPi.GPIO==0.7.1
spidev==3.6
SQLAlchemy==2.0.37
typing-extensions==4.12.2
werkzeug==3.1.3
zipp==3.21.0
